Correção da disância de hamming: considerando 8 bits por byte e retornando valor relativo.

+++ TrabalhoCriptografia/TrabalhoCriptografia/AES.h
@@ -34,7 +34,7 @@ void decryptBlock(byte* data, byte** allKeys, int rounds, byte** toXor, byte* re
 
 void matrixTransposer(byte* data);
 long calculateOnBits(byte b);
-long calculateHammingDistance(byte * clearM, byte * criptoM, int arraySize);
+float calculateHammingDistance(byte * clearM, byte * criptoM, int arraySize);
 //void byteStuffer(byte * b, int size);
 
 void vigenereCipherEncryption(byte *bytes, byte *key);
